Set a darker theme Cleaned up the UI to be more friendly and be a single "options" button Added in Recipe page functionality -Deletion of Recipes and Items currently fails

+++ src/main/java/home/inventory/beans/InventoryBean.java
@@ -2,6 +2,7 @@ package home.inventory.beans;
 
 import home.inventory.beans.data.PantryItemLazyLoader;
 import home.inventory.entities.PantryItem;
+import home.inventory.repos.IngredientRepo;
 import java.io.Serializable;
 import javax.enterprise.context.SessionScoped;
 import javax.faces.application.FacesMessage;
@@ -10,8 +11,6 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import org.apache.deltaspike.jpa.api.transaction.Transactional;
 import home.inventory.repos.PantryItemRepo;
-import java.util.ArrayList;
-import java.util.List;
 import org.primefaces.model.LazyDataModel;
 
 /**
@@ -26,11 +25,13 @@ public class InventoryBean implements Serializable {
 
     @Inject
     private PantryItemRepo itemRepo;
+    @Inject
+    private IngredientRepo ingredientRepo;
     private String name;
     private double quantity;
     private String units;
     private double quantityModifier; //used to store how much should be added or removed from an item
-    private String selected;
+    private PantryItem selected;
     @Inject
     private PantryItemLazyLoader items;
 
@@ -55,7 +56,7 @@ public class InventoryBean implements Serializable {
     }
 
     private void modifyItemQuantity(double modifier) {
-        PantryItem item = itemRepo.findBy(selected);
+        PantryItem item = itemRepo.findBy(selected.getName());
         String action = "";
         String result = "Error";
         if (item != null) {
@@ -75,6 +76,7 @@ public class InventoryBean implements Serializable {
             action = "Unable to modify item quantity";
         }
         fctx().addMessage(null, new FacesMessage(result, action));
+        clear();
     }
 
     @Transactional
@@ -89,9 +91,22 @@ public class InventoryBean implements Serializable {
 
     @Transactional
     public void deleteItem() {
-        PantryItem item = itemRepo.findBy(selected);
-        //need some query to check that an item isn't part of a recipe
-        itemRepo.remove(item);
+        PantryItem item = itemRepo.findBy(selected.getName());
+        if(ingredientRepo.findAllByItemName(item.getName()).getResultList().isEmpty()){
+            itemRepo.remove(item);
+            clear();
+            fctx().addMessage(null, new FacesMessage("Success", "Deleted item"));
+        } else {
+            fctx().addMessage(null, new FacesMessage("Error", "Selected item has associated recipes. Please remove it from those recipes first"));
+        }
+    }
+    @Transactional
+    public void updateUnits() {
+        PantryItem item = itemRepo.findBy(selected.getName());
+        item.setUnits(units);
+        itemRepo.save(item);
+        clear();
+        fctx().addMessage(null, new FacesMessage("Success", "Updated units to "+units));
     }
 
     public String getName() {
@@ -126,11 +141,11 @@ public class InventoryBean implements Serializable {
         this.quantityModifier = quantityModifier;
     }
 
-    public String getSelected() {
+    public PantryItem getSelected() {
         return selected;
     }
 
-    public void setSelected(String selected) {
+    public void setSelected(PantryItem selected) {
         this.selected = selected;
     }
